refactoring: create LoginAdministratorAsUserUrlProviderInterface to allow the feature overriding
