+++ packages/frontend-api/src/Model/Security/LoginAdministratorAsUserUrlProvider.php
@@ -2,17 +2,17 @@
 
 declare(strict_types=1);
 
-namespace Shopsys\FrontendApiBundle\Twig;
+namespace Shopsys\FrontendApiBundle\Model\Security;
 
+use Override;
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
 use Shopsys\FrameworkBundle\Component\Router\DomainRouterFactory;
 use Shopsys\FrameworkBundle\Controller\Admin\LoginController;
 use Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser;
+use Shopsys\FrameworkBundle\Model\Security\LoginAdministratorAsUserUrlProvider as BaseLoginAdministratorAsUserUrlProvider;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
-use Twig\Extension\AbstractExtension;
-use Twig\TwigFunction;
 
-class LoginAsUserExtension extends AbstractExtension
+class LoginAdministratorAsUserUrlProvider extends BaseLoginAdministratorAsUserUrlProvider
 {
     /**
      * @param \Shopsys\FrameworkBundle\Component\Router\DomainRouterFactory $domainRouterFactory
@@ -23,19 +23,10 @@ class LoginAsUserExtension extends AbstractExtension
     }
 
     /**
-     * @return array
-     */
-    public function getFunctions(): array
-    {
-        return [
-            new TwigFunction('getSsoLoginAsCustomerUserUrl', $this->getSsoLoginAsCustomerUserUrl(...)),
-        ];
-    }
-
-    /**
      * @param \Shopsys\FrameworkBundle\Model\Customer\User\CustomerUser $customerUser
      * @return string
      */
+    #[Override]
     public function getSsoLoginAsCustomerUserUrl(CustomerUser $customerUser): string
     {
         $customerDomainRouter = $this->domainRouterFactory->getRouter($customerUser->getDomainId());
