+++ packages/frontend-api/src/Resources/translations/messages.en.po
@@ -7,6 +7,3 @@ msgstr ""
 msgid "Customer not found."
 msgstr ""
 
-msgid "Log in as user"
-msgstr ""
-
