+++ project-base/app/config/bundles.php
@@ -11,7 +11,6 @@ return [
     Presta\SitemapBundle\PrestaSitemapBundle::class => ['all' => true],
     Prezent\Doctrine\TranslatableBundle\PrezentDoctrineTranslatableBundle::class => ['all' => true],
     Shopsys\FormTypesBundle\ShopsysFormTypesBundle::class => ['all' => true],
-    Shopsys\FrontendApiBundle\ShopsysFrontendApiBundle::class => ['all' => true],
     Shopsys\MigrationBundle\ShopsysMigrationBundle::class => ['all' => true],
     Shopsys\ProductFeed\HeurekaBundle\ShopsysProductFeedHeurekaBundle::class => ['all' => true],
     Shopsys\ProductFeed\GoogleBundle\ShopsysProductFeedGoogleBundle::class => ['all' => true],
@@ -34,6 +33,7 @@ return [
     FOS\CKEditorBundle\FOSCKEditorBundle::class => ['all' => true],
     Fp\JsFormValidatorBundle\FpJsFormValidatorBundle::class => ['all' => true],
     Shopsys\FrameworkBundle\ShopsysFrameworkBundle::class => ['all' => true],
+    Shopsys\FrontendApiBundle\ShopsysFrontendApiBundle::class => ['all' => true],
     Symfony\Bundle\WebProfilerBundle\WebProfilerBundle::class => ['dev' => true],
     Symfony\Bundle\DebugBundle\DebugBundle::class => ['dev' => true],
     Symfony\WebpackEncoreBundle\WebpackEncoreBundle::class => ['all' => true],
