+++ upgrade-notes/backend_20240529_084428.md
@@ -1010,10 +1010,11 @@
     -       protected readonly LoginAsUserFacade $loginAsUserFacade,
             protected readonly DomainRouterFactory $domainRouterFactory,
             // ...
+    +       protected readonly LoginAdministratorAsUserUrlProvider $loginAdministratorAsUserUrlProvider,
     ```
     -   `LOGIN_AS_TOKEN_ID_PREFIX` constant was removed
     -   `loginAsUserAction()` method was removed, use `Shopsys\FrontendApiBundle\Controller\CustomerUserController::loginAsCustomerUserAction()` instead
-    -   `getSsoLoginAsCustomerUserUrl()` method was removed, use `Shopsys\FrontendApiBundle\Twig\LoginAsUserExtension::getSsoLoginAsCustomerUserUrl()` instead
+    -   `getSsoLoginAsCustomerUserUrl()` method was removed, use `Shopsys\FrontendApiBundle\Model\Security\LoginAdministratorAsUserUrlProvider::getSsoLoginAsCustomerUserUrl()` instead
 -   `Shopsys\FrameworkBundle\Model\Administrator\Administrator` class was changed:
     -   the class now implements `Scheb\TwoFactorBundle\Model\Email\TwoFactorInterface` and `Scheb\TwoFactorBundle\Model\Google\TwoFactorInterface`
     -   `checkRolesContainAdminRole()` method was removed
